+++ lib/cuke_modeler/adapters/gherkin_26_adapter.rb
@@ -1,0 +1,13 @@
+require_relative 'gherkin_20_adapter'
+
+
+module CukeModeler
+
+  # NOT A PART OF THE PUBLIC API
+  # An adapter that can convert the output of version 26.x of the *cucumber-gherkin* gem into input that is consumable
+  # by this gem.
+
+  class Gherkin26Adapter < Gherkin20Adapter
+
+  end
+end